Arreglado precio decimal

+++ database/migrations/2020_01_08_012803_create_products_table.php
@@ -20,7 +20,7 @@ class CreateProductsTable extends Migration
             $table->string('slug');
             $table->integer('category_id');
             $table->string('image');
-            $table->decimal('price', 11,2);
+            $table->integer('price');
             $table->integer('in_discount');
             $table->integer('discount');
             $table->text('content');
